+++ packages/composer/CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- composer releases a ES module bundle at `/dist/bundle/index.min.js`
+
+### Fixed
+
+- composer can render a single string node like `Hello 👋`
+
 ## [0.4.0] - 2022-11-21
 
 ### Added